Refractor internal error-handling
All errors are now wrapped with a locally defined context. That gives 2 main
benefits:

1. One can easily match on a specifc type of error internally and respond
differently. E.g a `ConnectionError` event could be matched on and the client
could wait a certain timeout before retrying.

2. It is easy to add new local errortypes, or wrap existing errortypes from
dependencies with extra information. This causes detailed backtraces which can
be used to quickly isolate errors.

What is left to do is to incorporate an external errortype from the
'datatypes'-crate, which will be sent across 'tarpc' to end users.